Move JsonExplicit to the only place it's used.
