+++ storage/src/main/java/tech/pegasys/teku/storage/server/metadata/JsonExplicit.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.serialization;
+package tech.pegasys.teku.storage.server.metadata;
 
 import com.fasterxml.jackson.annotation.JacksonAnnotationsInside;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
